# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API system for Trekkstay - Hotel Booking System",
        "title": "Trekkstay - Hotel Booking System API",
        "termsOfService": "https://swagger.io/",
        "contact": {
            "name": "Trekkstay Team",
            "url": "https://www.trekkstay.com",
            "email": "support@trekkstay.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v0.0.1"
    },
    "host": "175.41.168.200:8888",
    "basePath": "/api/v1",
    "paths": {
...
        "/hotel-room/create": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create new hotel room, requires authentication with owner role",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hotel Room"
                ],
                "summary": "Create new hotel room",
                "parameters": [
                    {
                        "description": "CreateHotelRoomReq JSON",
                        "name": "CreateHotelRoomReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.CreateHotelRoomReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/res.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/res.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/res.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/hotel/create": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create new hotel, requires authentication with owner role",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hotel"
                ],
                "summary": "Create new hotel",
                "parameters": [
                    {
                        "description": "CreateHotelReq JSON",
                        "name": "CreateHotelReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.CreateHotelReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/res.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/res.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/res.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/payment/create": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create new payment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Create new payment",
                "parameters": [
                    {
                        "description": "CreatePaymentReq JSON",
                        "name": "CreatePaymentReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.CreatePaymentReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/res.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/res.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/res.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payment/update": {
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Update payment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Update payment",
                "parameters": [
                    {
                        "description": "UpdatePaymentReq JSON",
                        "name": "UpdatePaymentReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.UpdatePaymentReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/res.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/res.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/reservation/cancel/{reservation_id}": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Cancel reservation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reservation"
                ],
                "summary": "Cancel reservation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation ID",
                        "name": "reservation_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/res.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/res.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "modules_hotel_api_model_req.MediaJSON": {
            "type": "object",
            "properties": {
                "urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "modules_hotel_room_api_model_req.MediaJSON": {
            "type": "object",
            "properties": {
                "urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "req.CoordinatesJSON": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                }
            }
        },
...
        "req.CreateHotelReq": {
            "type": "object",
            "required": [
                "address_detail",
                "check_in_time",
                "check_out_time",
                "description",
                "district_code",
                "email",
                "name",
                "phone",
                "province_code",
                "ward_code"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "x-order": "1"
                },
                "description": {
                    "type": "string",
                    "x-order": "10"
                },
                "facilities": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/req.HotelFacilitiesJSON"
                        }
                    ],
                    "x-order": "11"
                },
                "coordinates": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/req.CoordinatesJSON"
                        }
                    ],
                    "x-order": "12"
                },
                "videos": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/modules_hotel_api_model_req.MediaJSON"
                        }
                    ],
                    "x-order": "13"
                },
                "images": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/modules_hotel_api_model_req.MediaJSON"
                        }
                    ],
                    "x-order": "14"
                },
                "email": {
                    "type": "string",
                    "x-order": "2"
                },
                "phone": {
                    "type": "string",
                    "x-order": "3"
                },
                "check_in_time": {
                    "type": "string",
                    "x-order": "4"
                },
                "check_out_time": {
                    "type": "string",
                    "x-order": "5"
                },
                "province_code": {
                    "type": "string",
                    "x-order": "6"
                },
                "district_code": {
                    "type": "string",
                    "x-order": "7"
                },
                "ward_code": {
                    "type": "string",
                    "x-order": "8"
                },
                "address_detail": {
                    "type": "string",
                    "x-order": "9"
                }
            }
        },
        "req.CreateHotelRoomReq": {
            "type": "object",
            "required": [
                "description",
                "facilities",
                "hotel_id",
                "original_price",
                "quantity",
                "type"
            ],
            "properties": {
                "hotel_id": {
                    "type": "string",
                    "x-order": "1"
                },
                "type": {
                    "type": "string",
                    "x-order": "2"
                },
                "description": {
                    "type": "string",
                    "x-order": "3"
                },
                "quantity": {
                    "type": "integer",
                    "x-order": "4"
                },
                "discount_rate": {
                    "type": "integer",
                    "x-order": "5"
                },
                "original_price": {
                    "type": "integer",
                    "x-order": "6"
                },
                "videos": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/modules_hotel_room_api_model_req.MediaJSON"
                        }
                    ],
                    "x-order": "7"
                },
                "images": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/modules_hotel_room_api_model_req.MediaJSON"
                        }
                    ],
                    "x-order": "8"
                },
                "facilities": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/req.HotelRoomFacilitiesJSON"
                        }
                    ],
                    "x-order": "9"
                }
            }
        },
        "req.CreatePaymentReq": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 1000000
                },
                "method": {
                    "type": "string",
                    "example": "MOMO"
                },
                "reservation_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "PENDING"
                }
            }
        },
...
        "req.HotelFacilitiesJSON": {
            "type": "object",
            "properties": {
                "airport_transfer": {
                    "type": "boolean"
                },
                "conference_room": {
                    "type": "boolean"
                },
                "fitness_center": {
                    "type": "boolean"
                },
                "food_service": {
                    "type": "boolean"
                },
                "free_wifi": {
                    "type": "boolean"
                },
                "laundry_service": {
                    "type": "boolean"
                },
                "motor_bike_rental": {
                    "type": "boolean"
                },
                "parking_area": {
                    "type": "boolean"
                },
                "spa_service": {
                    "type": "boolean"
                },
                "swimming_pool": {
                    "type": "boolean"
                }
            }
        },
        "req.HotelRoomFacilitiesJSON": {
            "type": "object",
            "properties": {
                "air_conditioner": {
                    "type": "boolean"
                },
                "balcony": {
                    "description": "Default: false",
                    "type": "boolean"
                },
                "bath_tub": {
                    "type": "boolean"
                },
                "hair_dryer": {
                    "type": "boolean"
                },
                "kitchen": {
                    "type": "boolean"
                },
                "non_smoking": {
                    "type": "boolean"
                },
                "number_of_bed": {
                    "type": "integer"
                },
                "room_size": {
                    "type": "integer"
                },
                "shower": {
                    "type": "boolean"
                },
                "sleeps": {
                    "description": "Jsonb",
                    "allOf": [
                        {
                            "$ref": "#/definitions/req.SleepJSON"
                        }
                    ]
                },
                "slippers": {
                    "type": "boolean"
                },
                "television": {
                    "description": "Default: true",
                    "type": "boolean"
                },
                "view": {
                    "type": "string"
                }
            }
        },
...
        "req.SleepJSON": {
            "type": "object",
            "properties": {
                "adults": {
                    "type": "integer"
                },
                "children": {
                    "type": "integer"
                }
            }
        },
...
        "req.UpdatePaymentReq": {
            "type": "object",
            "properties": {
                "reservation_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "SUCCESS"
                }
            }
        },
...
        "res.ErrorResponse": {
            "type": "object",
            "properties": {
                "error_key": {
                    "type": "string"
                },
                "log": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "res.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}